% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_147_leap_year_calculator.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Check if a Year is a Leap Year}
\usage{
is_leap_year(year)
}
\arguments{
\item{year}{An integer representing the year to check.}
}
\value{
A logical value: \code{TRUE} if the year is a leap year,
\code{FALSE} otherwise.
}
\description{
Determines whether a given year is a leap year according to the Gregorian
calendar rules. A year is a leap year if it is divisible by 4, except for
century years which must be divisible by 400.
}
\examples{
is_leap_year(2024)  # TRUE
is_leap_year(1900)  # FALSE
is_leap_year(2000)  # TRUE

}
